requirements.txt
aniso8601==9.0.1
async-timeout==4.0.3
attrs==24.2.0
blinker==1.8.2
cachelib==0.9.0
cffi==1.17.1
click==8.1.7
cryptography==43.0.1
Flask==3.0.0
Flask-Caching==2.3.0
Flask-Cors==5.0.0
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jwt==1.3.1
MarkupSafe==2.1.5
pillow==10.4.0
pkgutil_resolve_name==1.3.10
pycparser==2.22
PyMySQL==1.1.1
pytz==2024.2
PyYAML==6.0.2
redis==5.0.2
referencing==0.35.1
rpds-py==0.20.0
SQLAlchemy==2.0.35
typing_extensions==4.12.2
Werkzeug==3.0.4
zipp==3.20.2
